userAppProject/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
curl==0.0.1
Flask==1.1.2
Flask-Limiter==1.4
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
flask-swagger-ui==3.36.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
limits==1.5.1
MarkupSafe==1.1.1
marshmallow==3.11.1
mistune==0.8.4
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
